:lipstick:style 一覧が見易いようにheaderとfooterを調整しました
