+++ front/src/app/components/Layouts/Footer.tsx
@@ -1,0 +1,8 @@
+export default function Footer() {
+
+  return(
+    <footer className="bg-white">
+      <div className="py-6"></div>
+    </footer>
+  );
+}
